refactor compilation utils
