+++ mathics/builtin/compile/types.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+
+from llvmlite import ir
+
+int_type = ir.IntType(64)
+real_type = ir.DoubleType()
+bool_type = ir.IntType(1)
+void_type = ir.VoidType()
